+++ src/main/java/xyz/tomorrowlearncamp/newsfeed/domain/comment/dto/CreateCommentRequestDto.java
@@ -5,15 +5,16 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
+import xyz.tomorrowlearncamp.newsfeed.global.etc.DtoErrorMessage;
 
 @Getter
 public class CreateCommentRequestDto {
 
-    @NotBlank(message = "댓글을 입력하세요")
-    @Size(min = 2, max = 1000)
+    @NotBlank(message = DtoErrorMessage.EMPTY_CONTENT)
+    @Size(min = 2, max = 1000, message = DtoErrorMessage.INVALID_CONTENT_LENGTH)
     private final String content;
 
-    @NotNull(message = "게시글 필수")
+    @NotNull(message = DtoErrorMessage.EMPTY_CONTENT)
     private final Long newsFeedId;
 
     private final Integer depth;
