+++ src/main/java/xyz/tomorrowlearncamp/newsfeed/domain/newsFeeds/dto/request/UpdateNewsFeedRequestDto.java
@@ -1,6 +1,5 @@
-package xyz.tomorrowlearncamp.newsfeed.domain.newsFeeds.dto.requestDto;
+package xyz.tomorrowlearncamp.newsfeed.domain.newsFeeds.dto.request;
 
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
