+++ src/main/java/xyz/tomorrowlearncamp/newsfeed/domain/user/dto/request/DeleteUsersRequestDto.java
@@ -6,13 +6,13 @@ import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import xyz.tomorrowlearncamp.newsfeed.global.etc.Const;
-import xyz.tomorrowlearncamp.newsfeed.global.etc.ErrorMessage;
+import xyz.tomorrowlearncamp.newsfeed.global.etc.DtoErrorMessage;
 
 @Getter
 @AllArgsConstructor
 public class DeleteUsersRequestDto {
-    @NotNull(message = ErrorMessage.EMPTY_CONTENT)
-    @Size(min = 8, max = 20, message = ErrorMessage.INVALID_PASSWORD_LENGTH)
-    @Pattern(regexp = Const.PASSWORD_REGEX, message = ErrorMessage.INVALID_PASSWORD_FORMAT)
+    @NotNull(message = DtoErrorMessage.EMPTY_CONTENT)
+    @Size(min = 8, max = 20, message = DtoErrorMessage.INVALID_PASSWORD_LENGTH)
+    @Pattern(regexp = Const.PASSWORD_REGEX, message = DtoErrorMessage.INVALID_PASSWORD_FORMAT)
     private final String password;
 }
