+++ src/main/java/xyz/tomorrowlearncamp/newsfeed/domain/user/dto/request/UpdateUsersRequestDto.java
@@ -8,32 +8,30 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import xyz.tomorrowlearncamp.newsfeed.domain.user.enums.Gender;
 import xyz.tomorrowlearncamp.newsfeed.global.etc.Const;
-import xyz.tomorrowlearncamp.newsfeed.global.etc.ErrorMessage;
+import xyz.tomorrowlearncamp.newsfeed.global.etc.DtoErrorMessage;
 
-import java.sql.Timestamp;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 @Getter
 @AllArgsConstructor
 public class UpdateUsersRequestDto {
     private final String username;
 
-    @NotNull(message = ErrorMessage.EMPTY_CONTENT)
-    @Pattern(regexp = Const.EMAIL_REGEX, message = ErrorMessage.INVALID_EMAIL_FORMAT)
-    @Email(message = ErrorMessage.INVALID_EMAIL_FORMAT)
-    @Size(min = 10, max = 30, message = ErrorMessage.INVALID_EMAIL_LENGTH)
+    @NotNull(message = DtoErrorMessage.EMPTY_CONTENT)
+    @Pattern(regexp = Const.EMAIL_REGEX, message = DtoErrorMessage.INVALID_EMAIL_FORMAT)
+    @Email(message = DtoErrorMessage.INVALID_EMAIL_FORMAT)
+    @Size(min = 10, max = 30, message = DtoErrorMessage.INVALID_EMAIL_LENGTH)
     private final String email;
 
-    @NotNull(message = ErrorMessage.EMPTY_CONTENT)
-    @Size(min = 8, max = 20, message = ErrorMessage.INVALID_PASSWORD_LENGTH)
-    @Pattern(regexp = Const.PASSWORD_REGEX, message = ErrorMessage.INVALID_PASSWORD_FORMAT)
+    @NotNull(message = DtoErrorMessage.EMPTY_CONTENT)
+    @Size(min = 8, max = 20, message = DtoErrorMessage.INVALID_PASSWORD_LENGTH)
+    @Pattern(regexp = Const.PASSWORD_REGEX, message = DtoErrorMessage.INVALID_PASSWORD_FORMAT)
     private final String password;
 
-    @NotNull(message = ErrorMessage.MISSING_GENDER)
+    @NotNull(message = DtoErrorMessage.MISSING_GENDER)
     private final Gender gender;
 
-    @NotNull(message = ErrorMessage.MISSING_BIRTHDATE)
+    @NotNull(message = DtoErrorMessage.MISSING_BIRTHDATE)
     private final LocalDate birthDate;
 
 }
